backend/requirements.txt
pydantic>=1.10,<2
fastapi==0.95.2
uvicorn==0.22.0
httpx>=0.24,<0.28
openai==1.12.0
yt-dlp
websocket-client>=1.6.0
apscheduler==3.10.4
python-dotenv==1.0.1
TikTokApi>=7.0
google-api-python-client>=2.100
apify-client>=1.6.0
